Parameter update (specificity, dropout, compliance, asymp frac, mild frac, gen time) (#119)
* Adjust parameters for testing specificity, dropout and compliance rates, fraction asymptomatic, mild fraction, and generation time distribution; adjust script for processing parameter dictionary so that whitespace is stripped from the parameter-type column before unique values are found

* Lower min number of cases to test in test_recursive_testing

Fix test_quarantine_household_on_symptoms to only look at index cases
from the end time. This is required because when we filter out
hospitalised people we only do it for the final time. Also remove
days_since_index/contact calcs which are no longer used.

* Fix hospital test test_interaction_type_representative

Missing ICU interactions due to the insufficient overall cases at the
end of the interaction. Increase the number of seed infections and
reduce the end_time to ensure a large number of cases and the end of the
test (hence ICU interactions).

Co-authored-by: roberthinch <robert.hinch@gmail.com>