Add instantaneous R calculation (#171)
* Bugfix - remove occupation network from params object once set up

If model is created multiple times from the same params object then we
get different answers. Problem with the Simulation Python interface, but
not the Model Python interface (so was missed in tests).

* Fix problem with n_tests in Python interface when tests are immediate

Add test to check the number of tests output is correct given the number
of prior day new infections

* Remove restriction that recursive testing cannot be immediate

Make sure that all scheduled tests are processed in one_time_step
(previously this was ensured by not allowing parameter choices which
would require this). Remove param check.

Add test to make sure all tests are now being processed at the right
time.

* Add calculation of the instantaneous R with confidence intervals

R is calculated using the known generation distribution and time-series
of new infections

A flat prior on R is used

* Go back to day 0 to get the seed infections

If either actual or expected infections is less than 1 then don't give
an estimate of R