Massive speed up by changing memory management of interactions
For each day we start with an array of interactions (reused every
days_of_interactions). This means when interactions are added they are
on a contiguous block of memory. Additionally they do not need to be
actively returned as we reset the index on the array of interactions to
0 and then simply reuse.

When additional interactions are required (for new networks), we add an
additional interaction block for each day.