Performance upgrades (#202)
* Run Python_c interface test on model with 50k instead of 1m (faster)

* Massive speed up by changing memory management of interactions

For each day we start with an array of interactions (reused every
days_of_interactions). This means when interactions are added they are
on a contiguous block of memory. Additionally they do not need to be
actively returned as we reset the index on the array of interactions to
0 and then simply reuse.

When additional interactions are required (for new networks), we add an
additional interaction block for each day.

* Memory improvements 170Mb->110Mb for 100k people/keep 1 day of interacts

Add trace_tokens and events in blocks, so if not using certain features
then no need to add

Remove the constant MAX_DAILY_INTERACTIONS_KEPT and use
params->days_of_interactions instead (so can set to 1)

Use short instead of int in most places to reduce memory footprint

* Explicit type conversion for enumarated parameters

Check the value type using a getter and apply to value before calling
the setter (same for enumerated params as for standard params already).

Removes 27 test warnings.

* Fix test warning in test_interventions::test_priority_testing

Use df.insert() instead of assigning directly to a new column (both work
but latter causes a warning)

* First cut of multi-region meta-model

Run each sub-population in a sepearte thread and step forward in-synch

At the end of each time-step allow cases to be seeded between across
region and update parameters by region

* Fix add_user_network on the R interface

Add example R scripts for example_101 and example_102

* Add ability to override rng_seed in the multi-region model

* Add example_vaccination.R and example_add_network.R

* Increase initial seed to 10 (prevent stochastic extinction)

* Fix test - increase population size (reduce stochastic variability)