Implement polarising cross-immunity
At the point of recovering from a infection, work out which other
strains you have immunity to

Immunity last for the same time for all strains

Cross-immunity is perfectly correlated (i.e. when working out
cross-immunity pick a single random variable to calculate whether you
have immunity to other strains)

Add test if introducing a new strain to a population with herd immunity
to strain1. If perfect cross-immunity then expect few new infections of
the second strain, if no cross-immunity then expect a wave of the same
size of the original one.