Cross-immunity infrastructure (#208)
Add non-perfect cross-immunity between strains using a cross-immunity matrix.

Cross-immunity is perfectly correlated (i.e. when working out
cross-immunity pick a single random variable to calculate whether you
have immunity to other strains)

Add test if introducing a new strain to a population with herd immunity
to strain1. If perfect cross-immunity then expect few new infections of
the second strain, if no cross-immunity then expect a wave of the same
size of the original one.

Add monotonicity test, as the cross-immunity level increases we expect the number of infections
with a new strain to decrease

Co-authored-by: nbaya <nikolasbaya@gmail.com>
Co-authored-by: Nik Baya <40802812+nikbaya@users.noreply.github.com>